Fix handling of RIL_REQUEST_ENTER_SIM_PIN
First of all, if the data contains two pointers, the second might still
be NULL, causing "(null)" to be passed as the new PIN (two arguments means
enter PUK and new PIN). This was seen happening on 4.1.1 (Jelly Bean).

Second, the +CPIN command returns just OK on success, not +CREG! +CREG
could be sent as an unsolicited response, as a successful +CPIN command
will make the ME operational and could cause the ME to register itself
with the network, but this is not mandatory!

The +CREG if sent will be picked up by the RILJ layer anyway.

Signed-off-by: Ithamar R. Adema <ithamar.adema@team-embedded.nl>